feat: rename keyboard.js to default
