+++ src/index.js
@@ -1,4 +1,4 @@
-import { Keyboard } from './keyboardClass';
+import { Keyboard } from './keyboard';
 
 window.addEventListener("DOMContentLoaded", () => {
   const keyboard = new Keyboard();
